+++ IPTVPlayer/libs/decimal.py
@@ -575,7 +575,7 @@ class Decimal(object):
             return self
 
         # From an integer
-        if isinstance(value, (int, long)):
+        if isinstance(value, (int, int)):
             if value >= 0:
                 self._sign = 0
             else:
@@ -608,7 +608,7 @@ class Decimal(object):
                                  'from list or tuple.  The list or tuple '
                                  'should have exactly three elements.')
             # process sign.  The isinstance test rejects floats
-            if not (isinstance(value[0], (int, long)) and value[0] in (0, 1)):
+            if not (isinstance(value[0], (int, int)) and value[0] in (0, 1)):
                 raise ValueError("Invalid sign.  The first value in the tuple "
                                  "should be an integer; either 0 for a "
                                  "positive number or 1 for a negative number.")
@@ -622,7 +622,7 @@ class Decimal(object):
                 # process and validate the digits in value[1]
                 digits = []
                 for digit in value[1]:
-                    if isinstance(digit, (int, long)) and 0 <= digit <= 9:
+                    if isinstance(digit, (int, int)) and 0 <= digit <= 9:
                         # skip leading zeros
                         if digits or digit != 0:
                             digits.append(digit)
@@ -635,7 +635,7 @@ class Decimal(object):
                     self._int = ''.join(map(str, digits))
                     self._exp = value[2]
                     self._is_special = True
-                elif isinstance(value[2], (int, long)):
+                elif isinstance(value[2], (int, int)):
                     # finite number: digits give the coefficient
                     self._int = ''.join(map(str, digits or [0]))
                     self._exp = value[2]
@@ -760,12 +760,12 @@ class Decimal(object):
             return -((-1)**self._sign)
 
     def __eq__(self, other):
-        if not isinstance(other, (Decimal, int, long)):
+        if not isinstance(other, (Decimal, int, int)):
             return NotImplemented
         return self.__cmp__(other) == 0
 
     def __ne__(self, other):
-        if not isinstance(other, (Decimal, int, long)):
+        if not isinstance(other, (Decimal, int, int)):
             return NotImplemented
         return self.__cmp__(other) != 0
 
@@ -1427,7 +1427,7 @@ class Decimal(object):
 
         Equivalent to long(int(self))
         """
-        return long(self.__int__())
+        return int(self.__int__())
 
     def _fix_nan(self, context):
         """Decapitate the payload of a NaN to fit the context"""
@@ -4803,9 +4803,9 @@ def _ilog(x, M, L = 8):
     y = x-M
     # argument reduction; R = number of reductions performed
     R = 0
-    while (R <= L and long(abs(y)) << L-R >= M or
+    while (R <= L and int(abs(y)) << L-R >= M or
            R > L and abs(y) >> R-L >= M):
-        y = _div_nearest(long(M*y) << 1,
+        y = _div_nearest(int(M*y) << 1,
                          M + _sqrt_nearest(M*(M+_rshift_nearest(y, R)), M))
         R += 1
 
@@ -4954,18 +4954,18 @@ def _iexp(x, M, L=8):
     # expm1(z/2**(R-1)), ... , exp(z/2), exp(z).
 
     # Find R such that x/2**R/M <= 2**-L
-    R = _nbits((long(x)<<L)//M)
+    R = _nbits((int(x)<<L)//M)
 
     # Taylor series.  (2**L)**T > M
     T = -int(-10*len(str(M))//(3*L))
     y = _div_nearest(x, T)
-    Mshift = long(M)<<R
+    Mshift = int(M)<<R
     for i in xrange(T-1, 0, -1):
         y = _div_nearest(x*(Mshift + y), Mshift * i)
 
     # Expansion
     for k in xrange(R-1, -1, -1):
-        Mshift = long(M)<<(k+2)
+        Mshift = int(M)<<(k+2)
         y = _div_nearest(y*(y+Mshift), Mshift)
 
     return M+y
@@ -5066,7 +5066,7 @@ def _convert_other(other, raiseit=False):
     """
     if isinstance(other, Decimal):
         return other
-    if isinstance(other, (int, long)):
+    if isinstance(other, (int, int)):
         return Decimal(other)
     if raiseit:
         raise TypeError("Unable to convert %s to Decimal" % other)
